ENH: added casts for printing size_t in alloc_hook

+++ debugger/mi/alloc_hook.cpp
@@ -80,7 +80,7 @@ void* malloc(size_t size)
     if (alloc_file && alloc_file != ALLOC_FILE_TMP_VALUE)
     {
         void* addr = malloc_orig(size);
-        fprintf(alloc_file, "malloc %p %d\n", addr, size);
+        fprintf(alloc_file, "malloc %p %d\n", addr, (unsigned long int) size);
         fflush(alloc_file);
 
         return addr;
@@ -103,7 +103,8 @@ void* calloc(size_t num, size_t size)
     if (alloc_file && alloc_file != ALLOC_FILE_TMP_VALUE)
     {
         void* addr = calloc_orig(num, size);
-        fprintf(alloc_file, "calloc %p %d\n", addr, num * size);
+        fprintf(alloc_file, "calloc %p %d\n", addr,
+                            (unsigned long int) num * size);
         fflush(alloc_file);
 
         return addr;
@@ -126,7 +127,8 @@ void* realloc(void* addr, size_t size)
     if (alloc_file && alloc_file != ALLOC_FILE_TMP_VALUE)
     {
         void* addr_new = realloc_orig(addr, size);
-        fprintf(alloc_file, "realloc %p %p %d\n", addr, addr_new, size);
+        fprintf(alloc_file, "realloc %p %p %d\n", addr, addr_new,
+                            (unsigned long int) size);
         fflush(alloc_file);
 
         return addr_new;
